first batch of detectors
